move users controller to api namespace
