+++ app/controllers/api/v0/pois_controller.rb
@@ -1,8 +1,6 @@
 module Api
   module V0
     class PoisController < Api::V0::BaseController
-      protect_from_forgery with: :null_session, if: Proc.new { |c| c.request.format == 'application/json' }
-
       attr_writer :member_mailer
 
       def index
