+++ app/views/api/v0/users/error.json.jbuilder
